requirements.txt
flask
langchain
langchain-community
langchain-huggingface
langchain-openai
faiss-cpu
huggingface_hub
python-dotenv
